pyproject.toml
[project]
name = "meihe-villa-backend"
version = "0.1.0"
description = "FastAPI backend for Meihe Villa - Taiwan Heritage Sites (台灣古蹟網站)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "pydantic-settings>=2.0.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "aioboto3>=12.0.0",
    "python-multipart>=0.0.6",
    "email-validator>=2.3.0",
    "bcrypt>=4.0.0,<4.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.26.0",
    "aiosqlite>=0.19.0",
    "ruff>=0.2.0",
]

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
